+++ R/visualize.R
@@ -38,36 +38,46 @@
 #'
 #' @examples
 #' visualize(diamonds, "price:cut", type = "dist", fillcol = "red")
-#' visualize(diamonds, "carat:cut", yvar = "price", type = "scatter",
-#'   pointcol = "blue", fun = c("mean", "median"), linecol = c("red","green"))
-#' visualize(diamonds, yvar = "price", xvar = c("cut","clarity"),
-#'   type = "bar", fun = "median")
-#' visualize(diamonds, yvar = "price", xvar = c("cut","clarity"),
-#'   type = "line", fun = "max")
-#' visualize(diamonds, yvar = "price", xvar = "carat", type = "scatter",
-#'           size = "table", custom = TRUE) + scale_size(range = c(1, 10), guide = "none")
+#' visualize(diamonds, "carat:cut",
+#'   yvar = "price", type = "scatter",
+#'   pointcol = "blue", fun = c("mean", "median"), linecol = c("red", "green")
+#' )
+#' visualize(diamonds,
+#'   yvar = "price", xvar = c("cut", "clarity"),
+#'   type = "bar", fun = "median"
+#' )
+#' visualize(diamonds,
+#'   yvar = "price", xvar = c("cut", "clarity"),
+#'   type = "line", fun = "max"
+#' )
+#' visualize(diamonds,
+#'   yvar = "price", xvar = "carat", type = "scatter",
+#'   size = "table", custom = TRUE
+#' ) + scale_size(range = c(1, 10), guide = "none")
 #' visualize(diamonds, yvar = "price", xvar = "carat", type = "scatter", custom = TRUE) +
 #'   labs(title = "A scatterplot", x = "price in $")
 #' visualize(diamonds, xvar = "price:carat", custom = TRUE) %>%
 #'   wrap_plots(ncol = 2) + plot_annotation(title = "Histograms")
-#' visualize(diamonds, xvar = "cut", yvar = "price", type = "bar",
-#'   facet_row = "cut", fill = "cut")
+#' visualize(diamonds,
+#'   xvar = "cut", yvar = "price", type = "bar",
+#'   facet_row = "cut", fill = "cut"
+#' )
+#'
+#' @importFrom rlang .data
+#' @importFrom stats density
 #'
 #' @export
-visualize <- function(
-  dataset, xvar, yvar = "", comby = FALSE, combx = FALSE,
-  type = ifelse(radiant.data::is_empty(yvar), "dist", "scatter"), nrobs = -1,
-  facet_row = ".", facet_col = ".", color = "none", fill = "none",
-  size = "none", fillcol = "blue", linecol = "black", pointcol = "black",
-  bins = 10, smooth = 1, fun = "mean", check = "", axes = "",
-  alpha = 0.5, theme = "theme_gray", base_size = 11, base_family = "",
-  labs = list(), xlim = NULL, ylim = NULL, data_filter = "",
-  shiny = FALSE, custom = FALSE, envir = parent.frame()
-) {
-
+visualize <- function(dataset, xvar, yvar = "", comby = FALSE, combx = FALSE,
+                      type = ifelse(radiant.data::is_empty(yvar), "dist", "scatter"), nrobs = -1,
+                      facet_row = ".", facet_col = ".", color = "none", fill = "none",
+                      size = "none", fillcol = "blue", linecol = "black", pointcol = "black",
+                      bins = 10, smooth = 1, fun = "mean", check = "", axes = "",
+                      alpha = 0.5, theme = "theme_gray", base_size = 11, base_family = "",
+                      labs = list(), xlim = NULL, ylim = NULL, data_filter = "",
+                      shiny = FALSE, custom = FALSE, envir = parent.frame()) {
   if (missing(xvar) && type == "box") {
     xvar <- yvar
-    type = "box-single"
+    type <- "box-single"
     if (comby) {
       comby <- FALSE
       combx <- TRUE
@@ -147,7 +157,7 @@ visualize <- function(
   }
 
   ## get class
-  dc <- get_class(dataset)
+  dc <- dc_org <- get_class(dataset)
 
   ## if : is used to specify a range of variables
   if (length(vars) < ncol(dataset)) {
@@ -175,9 +185,9 @@ visualize <- function(
     dc <- get_class(dataset)
   }
 
-  if (type == "bar") {
+  if (type %in% c("bar", "line")) {
     if (any(xvar %in% yvar)) {
-      return("Cannot create a bar-chart if an X-variable is also included as a Y-variable")
+      return("Cannot create a bar or line chart if an X-variable is also included as a Y-variable")
     }
   } else if (type == "box") {
     if (any(xvar %in% yvar)) {
@@ -186,7 +196,7 @@ visualize <- function(
   }
 
   ## Determine if you want to use the first level of factor or not
-  if (type == "bar") {
+  if (type %in% c("bar", "line")) {
     isFctY <- "factor" == dc & names(dc) %in% yvar
     if (sum(isFctY)) {
       levs_org <- sapply(dataset[, isFctY, drop = FALSE], function(x) levels(x)[1])
@@ -220,7 +230,10 @@ visualize <- function(
       } else if (fun %in% c("median", "min", "max", "p01", "p025", "p05", "p10", "p25", "p50", "p75", "p90", "p95", "p975", "p99", "skew", "kurtosi")) {
         mfun <- fixer
       } else {
-        mfun <- function(x) { levs <<- c(levs, NA); x}
+        mfun <- function(x) {
+          levs <<- c(levs, NA)
+          x
+        }
       }
 
       dataset[, isFctY] <- select(dataset, which(isFctY)) %>%
@@ -259,13 +272,23 @@ visualize <- function(
 
   ## combining Y-variables if needed
   if (comby && length(yvar) > 1) {
-    if (any(xvar %in% yvar) && type != "box-single") return("X-variables cannot be part of Y-variables when combining Y-variables")
-    if (!radiant.data::is_empty(color, "none")) return("Cannot use Color when combining Y-variables")
-    if (!radiant.data::is_empty(fill, "none")) return("Cannot use Fill when combining Y-variables")
-    if (!radiant.data::is_empty(size, "none")) return("Cannot use Size when combining Y-variables")
-    if (facet_row %in% yvar || facet_col %in% yvar) return("Facet row or column variables cannot be part of\nY-variables when combining Y-variables")
+    if (any(xvar %in% yvar) && type != "box-single") {
+      return("X-variables cannot be part of Y-variables when combining Y-variables")
+    }
+    if (!radiant.data::is_empty(color, "none")) {
+      return("Cannot use Color when combining Y-variables")
+    }
+    if (!radiant.data::is_empty(fill, "none")) {
+      return("Cannot use Fill when combining Y-variables")
+    }
+    if (!radiant.data::is_empty(size, "none")) {
+      return("Cannot use Size when combining Y-variables")
+    }
+    if (facet_row %in% yvar || facet_col %in% yvar) {
+      return("Facet row or column variables cannot be part of\nY-variables when combining Y-variables")
+    }
 
-    dataset <- gather(dataset, "yvar", "values", !! yvar, factor_key = TRUE)
+    dataset <- gather(dataset, "yvar", "values", !!yvar, factor_key = TRUE)
     yvar <- "values"
     byvar <- if (is.null(byvar)) "yvar" else c("yvar", byvar)
     color <- fill <- "yvar"
@@ -275,11 +298,17 @@ visualize <- function(
 
   ## combining X-variables if needed
   if (combx && length(xvar) > 1) {
-    if (!radiant.data::is_empty(fill, "none")) return("Cannot use Fill when combining X-variables")
-    if (facet_row %in% xvar || facet_col %in% xvar) return("Facet row or column variables cannot be part of\nX-variables when combining Y-variables")
-    if (any(!get_class(select_at(dataset, .vars = xvar)) %in% c("numeric", "integer"))) return("Cannot combine plots for non-numeric variables")
+    if (!radiant.data::is_empty(fill, "none")) {
+      return("Cannot use Fill when combining X-variables")
+    }
+    if (facet_row %in% xvar || facet_col %in% xvar) {
+      return("Facet row or column variables cannot be part of\nX-variables when combining Y-variables")
+    }
+    if (any(!get_class(select_at(dataset, .vars = xvar)) %in% c("numeric", "integer"))) {
+      return("Cannot combine plots for non-numeric variables")
+    }
 
-    dataset <- gather(dataset, "xvar", "values", !! xvar, factor_key = TRUE)
+    dataset <- gather(dataset, "xvar", "values", !!xvar, factor_key = TRUE)
     xvar <- "values"
     byvar <- if (is.null(byvar)) "xvar" else c("xvar", byvar)
     color <- fill <- "xvar"
@@ -300,19 +329,21 @@ visualize <- function(
       hist_par <- list(alpha = alpha, position = "stack")
       if (combx) hist_par[["position"]] <- "identity"
       if (fill == "none") hist_par[["fill"]] <- fillcol
-      plot_list[[i]] <- ggplot(dataset, aes_string(x = i))
+      plot_list[[i]] <- ggplot(dataset, aes(x = .data[[i]]))
       if ("density" %in% axes && !"factor" %in% dc[i]) {
-        hist_par <- c(list(aes(y = ..density..)), hist_par)
-        plot_list[[i]] <- plot_list[[i]] + geom_density(adjust = smooth, color = linecol, size = .5)
+        hist_par <- c(list(aes(y = after_stat(density))), hist_par)
+        plot_list[[i]] <- plot_list[[i]] + geom_density(adjust = smooth, color = linecol, linewidth = .5)
       }
       if ("factor" %in% dc[i]) {
         plot_fun <- get("geom_bar")
         if ("log_x" %in% axes) axes <- sub("log_x", "", axes)
       } else {
         plot_fun <- get("geom_histogram")
-        hist_par[["binwidth"]] <- select_at(dataset, .vars = i) %>% range() %>% {
-          diff(.) / (bins - 1)
-        }
+        hist_par[["binwidth"]] <- select_at(dataset, .vars = i) %>%
+          range() %>%
+          {
+            diff(.) / (bins - 1)
+          }
       }
 
       plot_list[[i]] <- plot_list[[i]] + do.call(plot_fun, hist_par)
@@ -320,7 +351,7 @@ visualize <- function(
     }
   } else if (type == "density") {
     for (i in xvar) {
-      plot_list[[i]] <- ggplot(dataset, aes_string(x = i)) +
+      plot_list[[i]] <- ggplot(dataset, aes(x = .data[[i]])) +
         if (fill == "none") {
           geom_density(adjust = smooth, color = linecol, fill = fillcol, alpha = alpha, size = 1)
         } else {
@@ -350,7 +381,7 @@ visualize <- function(
       if ("log_x" %in% axes && dc[i] == "factor") axes <- sub("log_x", "", axes)
 
       for (j in yvar) {
-        plot_list[[itt]] <- ggplot(dataset, aes_string(x = i, y = j)) + gs
+        plot_list[[itt]] <- ggplot(dataset, aes(x = .data[[i]], y = .data[[j]])) + gs
 
         if ("log_x" %in% axes) plot_list[[itt]] <- plot_list[[itt]] + xlab(paste("log", i))
         if ("log_y" %in% axes) plot_list[[itt]] <- plot_list[[itt]] + ylab(paste("log", j))
@@ -450,7 +481,7 @@ visualize <- function(
       interpolate <- ifelse("interpolate" %in% check, TRUE, FALSE)
 
       for (j in yvar) {
-        plot_list[[itt]] <- ggplot(dataset, aes_string(x = i, y = j, fill = fill)) +
+        plot_list[[itt]] <- ggplot(dataset, aes(x = .data[[i]], y = .data[[j]], fill = .data[[fill]])) +
           geom_raster(interpolate = interpolate)
 
         if ("log_x" %in% axes) plot_list[[itt]] <- plot_list[[itt]] + xlab(paste("log", i))
@@ -465,7 +496,7 @@ visualize <- function(
       for (j in yvar) {
         flab <- ""
         if (color == "none") {
-          if (dc[i] %in% c("factor", "date")) {
+          if (dc[i] %in% c("factor", "date") || dc_org[j] == "factor") {
             tbv <- if (is.null(byvar)) i else c(i, byvar)
             tmp <- dataset %>%
               group_by_at(.vars = tbv) %>%
@@ -473,13 +504,15 @@ visualize <- function(
               na.omit() %>%
               summarise_all(fun)
             colnames(tmp)[ncol(tmp)] <- j
-            plot_list[[itt]] <- ggplot(tmp, aes_string(x = i, y = j)) + geom_line(aes(group = 1), color = linecol)
+            plot_list[[itt]] <- ggplot(tmp, aes(x = .data[[i]], y = .data[[j]])) +
+              geom_line(aes(group = 1), color = linecol)
             if (nrow(tmp) < 101) plot_list[[itt]] <- plot_list[[itt]] + geom_point(color = pointcol)
           } else {
-            plot_list[[itt]] <- ggplot(dataset, aes_string(x = i, y = j)) + geom_line(color = linecol)
+            plot_list[[itt]] <- ggplot(dataset, aes(x = .data[[i]], y = .data[[j]])) +
+              geom_line(color = linecol)
           }
         } else {
-          if (dc[i] %in% c("factor", "date")) {
+          if (dc[i] %in% c("factor", "date") || dc_org[j] == "factor") {
             tbv <- if (is.null(byvar)) i else unique(c(i, byvar))
             tmp <- dataset %>%
               group_by_at(.vars = tbv) %>%
@@ -487,25 +520,34 @@ visualize <- function(
               na.omit() %>%
               summarise_all(fun)
             colnames(tmp)[ncol(tmp)] <- j
-            plot_list[[itt]] <- ggplot(tmp, aes_string(x = i, y = j, color = color, group = color)) + geom_line()
+            plot_list[[itt]] <- ggplot(tmp, aes(x = .data[[i]], y = .data[[j]], color = .data[[color]], group = .data[[color]])) +
+              geom_line()
             if (nrow(tmp) < 101) plot_list[[itt]] <- plot_list[[itt]] + geom_point()
           } else {
-            plot_list[[itt]] <- ggplot(dataset, aes_string(x = i, y = j, color = color, group = color)) + geom_line()
+            plot_list[[itt]] <- ggplot(dataset, aes(x = .data[[i]], y = .data[[j]], color = .data[[color]], group = .data[[color]])) +
+              geom_line()
           }
         }
         if ("log_x" %in% axes) plot_list[[itt]] <- plot_list[[itt]] + xlab(paste("log", i))
         if ("log_y" %in% axes) plot_list[[itt]] <- plot_list[[itt]] + ylab(paste("log", j))
-        if (dc[i] %in% c("factor", "date") && nrow(tmp) < nrow(dataset)) {
-          plot_list[[itt]]$labels$y %<>% paste0(., " (", fun, ")")
+        if ((dc[i] %in% c("factor", "date") || dc_org[j] == "factor") && nrow(tmp) < nrow(dataset)) {
+          if (exists("levs")) {
+            if (j %in% names(levs) && !is.na(levs[j])) {
+              plot_list[[itt]]$labels$y %<>% paste0(., " (", fun, " {", levs[j], "})")
+            } else {
+              plot_list[[itt]]$labels$y %<>% paste0(., " (", fun, ")")
+            }
+          } else {
+            plot_list[[itt]]$labels$y %<>% paste0(., " (", fun, ")")
+          }
         }
-
         itt <- itt + 1
       }
     }
   } else if (type == "bar") {
     itt <- 1
     for (i in xvar) {
-      if (!"factor" %in% dc[i]) dataset[[i]] %<>% as_factor
+      if (!"factor" %in% dc[i]) dataset[[i]] %<>% as_factor()
       if ("log_x" %in% axes) axes <- sub("log_x", "", axes)
       for (j in yvar) {
         tbv <- if (is.null(byvar)) i else c(i, byvar)
@@ -525,7 +567,7 @@ visualize <- function(
           tmp[[i]] %<>% factor(., levels = unique(.))
         }
 
-        plot_list[[itt]] <- ggplot(tmp, aes_string(x = i, y = j)) + {
+        plot_list[[itt]] <- ggplot(tmp, aes(x = .data[[i]], y = .data[[j]])) + {
           if (fill == "none") {
             geom_bar(stat = "identity", position = "dodge", alpha = alpha, fill = fillcol)
           } else {
@@ -560,10 +602,10 @@ visualize <- function(
       if (!"factor" %in% dc[i]) dataset[[i]] %<>% as_factor
       for (j in yvar) {
         if (color == "none") {
-          plot_list[[itt]] <- ggplot(dataset, aes_string(x = i, y = j)) +
+          plot_list[[itt]] <- ggplot(dataset, aes(x = .data[[i]], y = .data[[j]])) +
             geom_boxplot(alpha = alpha, fill = fillcol, outlier.color = pointcol, color = linecol)
         } else {
-          plot_list[[itt]] <- ggplot(dataset, aes_string(x = i, y = j, fill = color)) +
+          plot_list[[itt]] <- ggplot(dataset, aes(x = .data[[i]], y = .data[[j]], fill = .data[[color]])) +
             geom_boxplot(alpha = alpha)
         }
 
@@ -579,22 +621,23 @@ visualize <- function(
   } else if (type == "box-single") {
     itt <- 1
     for (i in xvar) {
-        if (color == "none") {
-          plot_list[[itt]] <- dataset %>% ggplot(aes(x = "", y = .data[[i]])) +
-            geom_boxplot(alpha = alpha, fill = fillcol, outlier.color = pointcol, color = linecol) +
-            scale_x_discrete(labels = NULL, breaks = NULL) + labs(x = "")
-        } else {
-          plot_list[[itt]] <- dataset %>% ggplot(aes(x = "", y = .data[[i]], fill = color)) +
-            geom_boxplot(alpha = alpha)
-        }
+      if (color == "none") {
+        plot_list[[itt]] <- dataset %>% ggplot(aes(x = "", y = .data[[i]])) +
+          geom_boxplot(alpha = alpha, fill = fillcol, outlier.color = pointcol, color = linecol) +
+          scale_x_discrete(labels = NULL, breaks = NULL) +
+          labs(x = "")
+      } else {
+        plot_list[[itt]] <- dataset %>% ggplot(aes(x = "", y = .data[[i]], fill = color)) +
+          geom_boxplot(alpha = alpha)
+      }
 
-        if (!custom && (color == "none" || color == i)) {
-          plot_list[[itt]] <- plot_list[[itt]] + theme(legend.position = "none")
-        }
+      if (!custom && (color == "none" || color == i)) {
+        plot_list[[itt]] <- plot_list[[itt]] + theme(legend.position = "none")
+      }
 
-        if ("log_y" %in% axes) plot_list[[itt]] <- plot_list[[itt]] + ylab(paste("log", i))
+      if ("log_y" %in% axes) plot_list[[itt]] <- plot_list[[itt]] + ylab(paste("log", i))
 
-        itt <- itt + 1
+      itt <- itt + 1
     }
   }
 
@@ -607,41 +650,49 @@ visualize <- function(
     }
     scl <- if ("scale_y" %in% axes) "free_y" else "fixed"
     facet_fun <- if (facet_row == ".") facet_wrap else facet_grid
-    for (i in 1:length(plot_list))
+    for (i in 1:length(plot_list)) {
       plot_list[[i]] <- plot_list[[i]] + facet_fun(as.formula(facets), scales = scl)
+    }
   }
 
   if (color != "none") {
-    for (i in 1:length(plot_list))
-      plot_list[[i]] <- plot_list[[i]] + aes_string(color = color)
+    for (i in 1:length(plot_list)) {
+      plot_list[[i]] <- plot_list[[i]] + aes(color = .data[[color]])
+    }
   }
 
   if (size != "none") {
-    for (i in 1:length(plot_list))
-      plot_list[[i]] <- plot_list[[i]] + aes_string(size = size)
+    for (i in 1:length(plot_list)) {
+      plot_list[[i]] <- plot_list[[i]] + aes(size = .data[[size]])
+    }
   }
 
   if (fill != "none") {
-    for (i in 1:length(plot_list))
-      plot_list[[i]] <- plot_list[[i]] + aes_string(fill = fill)
+    for (i in 1:length(plot_list)) {
+      plot_list[[i]] <- plot_list[[i]] + aes(fill = .data[[fill]])
+    }
   }
 
   if ((length(xlim) == 2 && is.numeric(xlim)) &&
-      (length(ylim) == 2 && is.numeric(ylim))) {
-    for (i in 1:length(plot_list))
+    (length(ylim) == 2 && is.numeric(ylim))) {
+    for (i in 1:length(plot_list)) {
       plot_list[[i]] <- plot_list[[i]] + coord_cartesian(xlim = xlim, ylim = ylim)
+    }
   } else if (length(xlim) == 2 && is.numeric(xlim)) {
-    for (i in 1:length(plot_list))
+    for (i in 1:length(plot_list)) {
       plot_list[[i]] <- plot_list[[i]] + coord_cartesian(xlim = xlim)
+    }
   } else if (length(ylim) == 2 && is.numeric(ylim)) {
-    for (i in 1:length(plot_list))
+    for (i in 1:length(plot_list)) {
       plot_list[[i]] <- plot_list[[i]] + coord_cartesian(ylim = ylim)
+    }
   }
 
   if ("jitter" %in% check) {
-    for (i in 1:length(plot_list))
+    for (i in 1:length(plot_list)) {
       plot_list[[i]] <- plot_list[[i]] +
         geom_jitter(alpha = alpha, position = position_jitter(width = 0.4, height = 0.0))
+    }
   }
 
   if ("line" %in% check) {
@@ -687,7 +738,7 @@ visualize <- function(
     if (length(plot_list) == 1) plot_list[[1]] else plot_list
   } else {
     patchwork::wrap_plots(plot_list, ncol = min(length(plot_list), 2)) %>%
-      {if (shiny) . else print(.)}
+      (function(x) if (isTRUE(shiny)) x else print(x))
   }
 }
 
@@ -704,13 +755,15 @@ visualize <- function(
 #' qscatter(diamonds, "price", "carat")
 #' qscatter(titanic, "age", "survived")
 #'
+#' @importFrom rlang .data
+#'
 #' @export
 qscatter <- function(dataset, xvar, yvar, lev = "", fun = "mean", bins = 20) {
   if (is.character(dataset[[yvar]])) {
     dataset <- mutate_at(dataset, .vars = yvar, .funs = as.factor)
   }
   if (is.factor(dataset[[yvar]])) {
-    if (radiant.data::is_empty(lev)) lev <- levels(pull(dataset, !! yvar))[1]
+    if (radiant.data::is_empty(lev)) lev <- levels(pull(dataset, !!yvar))[1]
     dataset <- mutate_at(dataset, .vars = yvar, .funs = function(y) as.integer(y == lev))
     lev <- paste0(" {", lev, "}")
   } else {
@@ -719,7 +772,7 @@ qscatter <- function(dataset, xvar, yvar, lev = "", fun = "mean", bins = 20) {
   mutate_at(dataset, .vars = xvar, .funs = list(bins = ~ radiant.data::xtile(., bins))) %>%
     group_by(bins) %>%
     summarize_at(.vars = c(xvar, yvar), .funs = fun) %>%
-    ggplot(aes_string(x = xvar, y = yvar)) +
+    ggplot(aes(x = .data[[xvar]], y = .data[[yvar]])) +
     geom_point() +
     labs(y = paste0(yvar, " (", fun, lev, ")"))
 }